requirements.txt
streamlit
pydantic
google-generativeai
Pillow
pandas